feat: refactor menu to a separate file
